Flag any HANA as HANA Cloud
ATTENTION! This patch/hack makes the plugin only work against HANA Cloud
and thus must not be merged into the main branch!

This a very hacky workaround for the fact that some HANA Cloud instances
return an empty string as value for 'build_branch'. These instances are
then considered to be HANA on-premise instances because the build_branch
does not contain 'CE' (see code in HanaUtils.isCloudEditon()) which then
leads to the wrong SQL being sent in
HanaConnectionManager.loadWorkspaceMetadata().

Obviously this patch/hack makes the plugin only work against HANA Cloud
and thus must not be merged into the main branch!